requirements.txt
streamlit
pandas
geopandas
rasterio
matplotlib
seaborn
Pillow
openpyxl
scientific
shapely
scipy
numpy
requests
